instawork/requirements.txt
asgiref==3.3.1
Django==3.1.7
djangorestframework==3.12.2
pytz==2021.1
sqlparse==0.4.1
mysqlclient==1.4.6

